cleaner code to tes
